requirements.txt
flask
scikit-learn
pandas
numpy
joblib
groq
python-dotenv
